requirements.txt
numpy
pylint
pydocstyle
pycodestyle
mypy
pillow
matplotlib